move reminder to feature module
